iciar2020/requirements.txt
argparse
argparse_dataclass
cv2
efficientnet_pytorch
matplotlib
mpl_toolkits
numpy
pandas
pillow
scipy
seaborn
torch
torchvision

ietk
simplepytorch
